+++ .changeset/smooth-flowers-joke.md
@@ -1,0 +1,5 @@
+---
+"@bigcommerce/catalyst-core": patch
+---
+
+Applied streamable pattern to Cart.
